[#549] Rename package: local.storages > local.preferences
