+++ sample-compose/data/src/main/java/co/nimblehq/sample/compose/data/local/preferences/EncryptedSharedPreferences.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.sample.compose.data.local.storages
+package co.nimblehq.sample.compose.data.local.preferences
 
 import android.content.Context
 import androidx.security.crypto.EncryptedSharedPreferences
