+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/local/preferences/BaseSharedPreferences.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.sample.compose.data.local.storages
+package co.nimblehq.template.compose.data.local.preferences
 
 import android.content.SharedPreferences
