+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/local/preferences/EncryptedSharedPreferences.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.template.compose.data.local.storages
+package co.nimblehq.template.compose.data.local.preferences
 
 import android.content.Context
 import androidx.security.crypto.EncryptedSharedPreferences
